test(runtime): move page_object unit tests to sibling file, exclude from coverage
codecov/patch was failing (~54% < 80% target) because the inline #[cfg(test)]
module in src/page_object.rs hand-rolls MockDriver/MockElement doubles — ~30
`unimplemented!()` trait stubs that never execute. Living under src/, codecov
counted them as uncovered patch lines, even though the production logic they
test (execute_document_action, matcher-after-action ordering) is ~93% covered.

Relocate the whole test module to page_object_tests.rs, included via
`#[cfg(test)] #[path = ...] mod tests;` so it keeps private access to the
items under test, and add the file to codecov.yml's ignore list — mirroring
the existing exclusion of utam-runtime/tests/**. Test scaffolding shouldn't
be measured for coverage. No production code changes; all 33 lib tests pass.
